Split schemes to avoid circular import
